Refactor SidebarControl, fix up display, tests
